requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyfiglet==0.8.post1
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
robotframework==6.1.1
urllib3==2.0.4
Werkzeug==2.3.6
